Update: get_actions
